Balance editing progress

+++ app/Http/Controllers/StashController.php
@@ -3,10 +3,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StashShow;
 use App\Http\Requests\StashStore;
 use App\Http\Requests\StashUpdate;
-use App\Stash;
 use App\Http\Resources\StashResource;
+use App\Stash;
 use App\User;
 use Illuminate\Http\Request;
 
@@ -17,6 +18,16 @@ class StashController extends Controller
     public function __construct()
     {
         $this->middleware('auth:api');
+    }
+
+    protected function checkPermission(Stash $stash)
+    {
+        if (auth()->user()->id !== $stash->user_id) {
+            # TODO Translate
+            return response()->json(['error' => 'You can only edit your own stashes.'], 403);
+        }
+
+        return true;
     }
 
     public function index()
@@ -45,6 +56,11 @@ class StashController extends Controller
 
     public function show(Stash $stash)
     {
+        $perm = $this->checkPermission($stash);
+        if ($perm !== true) {
+            return $perm;
+        }
+
         return response()->json([
             'stash' => new StashResource($stash)
         ]);
@@ -52,9 +68,9 @@ class StashController extends Controller
 
     public function update(StashUpdate $request, Stash $stash)
     {
-        if (auth()->user()->id !== $stash->user_id) {
-            # TODO Translate
-            return response()->json(['error' => 'You can only edit your own stashes.'], 403);
+        $perm = $this->checkPermission($stash);
+        if ($perm !== true) {
+            return $perm;
         }
 
         $stash->update($request->validated());
@@ -66,16 +82,21 @@ class StashController extends Controller
 
     public function destroy(Stash $stash)
     {
+        $perm = $this->checkPermission($stash);
+        if ($perm !== true) {
+            return $perm;
+        }
+
         $id = $stash->id;
         $stash->delete();
 
-        return response()->json([ 'id' => $id ], 204);
+        return response()->json(['id' => $id], 204);
     }
 
     public function types()
     {
         return response()->json([
-            'types' => array_keys(Stash::STASH_TYPES),
+            'types' => array_keys(Stash::TYPES),
         ]);
     }
 }
